Remove core_utils
