+++ qiling/os/posix/posix.py
@@ -10,7 +10,6 @@ from unicorn.mips_const import *
 from unicorn.x86_const import *
 
 from qiling.const import *
-from qiling.core_utils import QlFileDes
 from qiling.os.os import QlOs
 from qiling.utils import *
 from qiling.exception import *
